refactors the john-hopkins models
