+++ covid/john_hopkins/__init__.py
@@ -1,0 +1,1 @@
+from .covid import Covid
